requirements.txt
# Core dependencies
requests==2.32.3
python-dotenv==1.0.1
# IANA timezones for zoneinfo (Windows); US/ET posting schedule in run_continuous_posts
tzdata>=2024.1
tenacity==9.0.0
Pillow==10.4.0

# AI and ML libraries
google-generativeai==0.8.5
google-cloud-aiplatform==1.38.1
openai==1.3.0

# News API and reputed sources (Google News, BBC/NPR/Reuters RSS)
gnews==0.2.1
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Facebook API
facebook-sdk==3.1.0

# Scheduling
schedule==1.2.0

# Additional utilities
urllib3==2.0.7

# Optional: remove text from generated images before overlay (REMOVE_TEXT_FROM_IMAGE). If missing, text removal is skipped.
opencv-python-headless>=4.8.0
easyocr>=1.7.0

# Image generation (imgen_feb / Z-Image-Turbo / FLUX) - install in the SAME env you run the script in
# TensorFlow (pulled in by transformers) needs protobuf 3.x with message/descriptor; 3.20.3 is known-good
protobuf>=3.20.0,<4.0.0
# If you see "cannot import name 'message' from 'google.protobuf'": run fix_protobuf.bat or:
#   .venv\Scripts\pip install --force-reinstall protobuf==3.20.3
# Z-Image (diffusers) + transformers require huggingface-hub>=0.34.0,<1.0 (transformers rejects 1.x)
torch>=2.0.0
torchvision>=0.15.0
# For CUDA 12.x: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
huggingface-hub>=0.24.0,<1.0
# FLUX (FLUX.1-schnell / FLUX.1-dev) - optional; models are gated (huggingface-cli login)
diffusers>=0.30.0
# Z-Image-Turbo (Tongyi-MAI) needs Qwen3Model; added in transformers 4.52.x
# diffusers/transformers need tokenizers in 0.22–0.23 range (0.20.x causes import errors)
tokenizers>=0.22.0,<=0.23.0
transformers>=4.52.0
accelerate>=0.33.0

# Optional: manual Cursor video — Kokoro narration + Whisper timed subtitles (see premium_voice_subtitles.py)
# pip install pykokoro soundfile openai-whisper
# Whisper pulls torch; use GPU for faster transcribe if available.
# Optional kinetic CSS captions (pycaps): pip install "pycaps[base] @ git+https://github.com/francozanardi/pycaps.git"